add more tests

+++ src/simulators.jl
@@ -562,9 +562,9 @@ end
 """
 Returns k almost equal patitions of n
 """
-@inline function equal_parts(n, k)
+@inline function equal_parts(n::Int, k::Int)
     ndiv = n ÷ k
     nrem = n % k
-    n_parts = ntuple(i -> (i <= nrem) ? ndiv+1 : ndiv, k)  # no allocation for k upto 10
+    n_parts = ntuple(i -> (i <= nrem) ? ndiv+1 : ndiv, k)
     return n_parts
 end
